Make a function that patches the client and call it.
